@@ -15,7 +15,8 @@
 from __future__ import annotations
 
 import itertools
-from typing import Hashable, Iterable, Iterator, Mapping, TypeVar
+from collections.abc import Hashable, Iterable, Iterator, Mapping
+from typing import TypeVar
 
 __all__ = ["CycleError", "lazy_stable_topo_sort"]
 
