@@ -14,7 +14,8 @@
 
 from __future__ import annotations
 
-from typing import Sequence, Union
+from collections.abc import Sequence
+from typing import Union
 
 __all__ = ["Python"]
 
