@@ -19,7 +19,8 @@ import importlib.util
 import json
 import os
 import sys
-from typing import TYPE_CHECKING, Sequence, TypeVar
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, TypeVar
 
 from colorlog.escape_codes import parse_colors
 
